User story 4 user roles

+++ app/controllers/wikis_controller.rb
@@ -1,14 +1,17 @@
 class WikisController < ApplicationController
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki).all
+    authorize @wikis
   end
 
   def show
-    @wiki = Wiki.find(params[:id])
+    @wiki = policy_scope(Wiki).find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
@@ -16,6 +19,7 @@ class WikisController < ApplicationController
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -27,13 +31,15 @@ class WikisController < ApplicationController
   end
 
   def edit
-    @wiki = Wiki.find(params[:id])
+    @wiki = policy_scope(Wiki).find(params[:id])
+    authorize @wiki
   end
 
   def update
-    @wiki = Wiki.find(params[:id])
+    @wiki = policy_scope(Wiki).find(params[:id])
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was updated."
@@ -45,7 +51,8 @@ class WikisController < ApplicationController
   end
 
   def destroy
-    @wiki = Wiki.find(params[:id])
+    @wiki = policy_scope(Wiki).find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "The wiki \"#{@wiki.title}\" has been successfully deleted"
